% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCAuth.R
\name{DCAuth}
\alias{DCAuth}
\title{Store Credentials for the DoubleClick Reporting API}
\usage{
DCAuth(client.id, client.secret, runwebserver = FALSE)
}
\arguments{
\item{client.secret}{}
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
DCAuth
}
\details{
Authorise and store credentials for the DoubleClick Reporting API
}
\examples{
\dontrun{
DCAuth("client.id", "client.secret")

}

}
