% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DCSetKey}
\alias{DCSetKey}
\title{Store Credentials for the DoubleClick Reporting API}
\usage{
DCSetKey(key)
}
\arguments{
\item{key}{}
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
DCSetKey
}
\details{
Set your server API key for the DoubleClick Reporting API
}
\examples{
\dontrun{
DCSetKey("your-key-here")

}
}

