% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RDoubleClick}
\alias{RDoubleClick}
\alias{RDoubleClick-package}
\title{R Client for DoubleClick Reporting API v1.43}
\description{
This package contains an "analyst's toolbox" of functions for accessing the DoubleClick Reporting API v1.3.
}
\details{
Package:  RDoubleClick \cr
Type:     Package \cr
Version:  0.1 \cr
Date:     2014-07-19 \cr
License:  MIT + file LICENSE \cr
}
\author{
Willem Paling
}
\references{
Official DoubleClick Reporting API documentation:

https://developers.google.com/doubleclick-advertisers/reporting/v1.3/

For support & bugs:

https://github.com/willempaling/RDoubleClick
}

