% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.list.R
\name{reports.list}
\alias{reports.list}
\title{List reports available in via the DoubleClick Reporting API}
\usage{
reports.list(profileId, results = 10, scope = "", sortField = "",
  sortOrder = "", fields = "", delay = 2)
}
\arguments{
\item{profileId}{The DFA profile ID - use userprofiles.list() to find yours}

\item{results}{The number of reports to return. Defaults to 10, can be increased to make multiple API requests and get more reports.}

\item{scope}{The scope that defines which results are returned, default is 'MINE'. Acceptable values are:
"ALL": All reports in account.
"MINE": My reports. (default)
"SHARED_WITH_ME": reports shared with me.}

\item{sortField}{The field by which to sort the list. 
Acceptable values are:
  "ID": Sort by file ID.
"LAST_MODIFIED_TIME": Sort by 'lastmodifiedAt' field. (default)}

\item{sortOrder}{Order of sorted results, default is 'DESCENDING', 'ASCENDING' is also valid.}

\item{fields}{Subset of fields to include in the report}

\item{delay}{Forced delay between API requests, to prevent errors}
}
\value{
File list (data table)
}
\description{
reports.list
}
\details{
List reports available in via the DoubleClick Reporting API
}
\examples{
\dontrun{
reports.list(1234567,results=100)

}

}
