% Generated by roxygen2 (4.0.1): do not edit by hand
\name{userprofiles.list}
\alias{userprofiles.list}
\title{Store Credentials for the DoubleClick Reporting API}
\usage{
userprofiles.list(client.id, client.secret)
}
\arguments{
\item{profileId}{}

\item{maxResults}{}

\item{pageToken}{}

\item{scope}{}

\item{sortField}{}

\item{sortOrder}{}

\item{fields}{}
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
userprofiles.list
}
\details{
Authorise and store credentials for the DoubleClick Reporting API
}
\examples{
\dontrun{
userprofiles.list("client.id", "client.secret")

}
}

