% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reports.run}
\alias{reports.run}
\title{Run previously defined report from DoubleClick Reporting API}
\usage{
reports.run(profileId, reportId)
}
\arguments{
\item{profileId}{The ID of the profile associated with the report}

\item{reportId}{The ID of the report}
}
\value{
data table containing report data
}
\description{
reports.run
}
\details{
Run previously defined report from DoubleClick Reporting API
}
\examples{
\dontrun{
reports.run(1234567,8765431)

}
}

