% Generated by roxygen2 (4.0.1): do not edit by hand
\name{userprofiles.list}
\alias{userprofiles.list}
\title{Lists available profiles for retrieving reports.}
\usage{
userprofiles.list(client.id, client.secret)
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
userprofiles.list
}
\details{
Lists available profiles for retrieving reports.
}
\examples{
\dontrun{
userprofiles.list()

}
}

