% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userprofiles.list.R
\name{userprofiles.list}
\alias{userprofiles.list}
\title{Lists available profiles for retrieving reports.}
\usage{
userprofiles.list(client.id, client.secret)
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
userprofiles.list
}
\details{
Lists available profiles for retrieving reports.
}
\examples{
\dontrun{
userprofiles.list()

}

}
